Refactor UI manager
Refactored `overlay_ui_manager.py` into smaller modules: `ui/header_manager.py`, `ui/control_panel_manager.py`, `ui/rf4s_control_manager.py`, and `ui/status_bar_manager.py`. This improves maintainability and code organization. Unused imports and files were removed.
