+++ pyside6_overlay/ui/__init__.py
@@ -6,8 +6,10 @@ Contains all user interface components and widgets
 
 from .panel_factory import PanelFactory
 from .workspace_manager import WorkspaceManager
+from .overlay_ui_manager import OverlayUIManager
 
 __all__ = [
     'PanelFactory',
-    'WorkspaceManager'
+    'WorkspaceManager',
+    'OverlayUIManager'
 ]
